+++ util/util.go
@@ -1,8 +1,9 @@
-//Package util implements other tool more.
+//Package util implements other tool more, like type cover, type value check.
 package util
 
 import (
 	"fmt"
+	"github.com/conero/uymas/str"
 	"math"
 	"reflect"
 	"time"
@@ -122,4 +123,40 @@ func ValueNull(value interface{}) bool {
 	}
 	v := reflect.ValueOf(value)
 	return v.IsZero()
+}
+
+// convert Struct field to by Map
+func StructToMap(value interface{}) map[string]interface{} {
+	rv := reflect.ValueOf(value)
+	if rv.Kind() == reflect.Struct {
+		rt := reflect.TypeOf(value)
+		vMap := map[string]interface{}{}
+		for i := 0; i < rv.NumField(); i++ {
+			field := rv.Field(i)
+			if field.Kind() != reflect.Func {
+				name := rt.Field(i).Name
+				vMap[name] = field.Interface()
+			}
+		}
+		return vMap
+	}
+	return nil
+}
+
+// convert Struct field to by Map and key is Lower style.
+func StructToMapLStyle(value interface{}) map[string]interface{} {
+	rv := reflect.ValueOf(value)
+	if rv.Kind() == reflect.Struct {
+		rt := reflect.TypeOf(value)
+		vMap := map[string]interface{}{}
+		for i := 0; i < rv.NumField(); i++ {
+			field := rv.Field(i)
+			if field.Kind() != reflect.Func {
+				name := rt.Field(i).Name
+				vMap[str.LowerStyle(name)] = field.Interface()
+			}
+		}
+		return vMap
+	}
+	return nil
 }